ref(billing): Cleanup billing details + payment method components
