+++ static/gsApp/components/creditCardEdit/legacyForm.tsx
@@ -41,13 +41,13 @@ export type SubmitData = {
 
 type Props = {
   /**
+   * budget term to use for fine print.
+   */
+  budgetTerm: string;
+  /**
    * Handle the card form submission.
    */
   onSubmit: (data: SubmitData) => void;
-  /**
-   * budget mode text for fine print, if any.
-   */
-  budgetModeText?: string;
   /**
    * Text for the submit button.
    */
@@ -91,8 +91,10 @@ type Props = {
  * Standalone credit card form that requires onSubmit to be handled
  * by the parent. This allows us to reuse the same form for both paymentintent, setupintent
  * and classic card flows.
+ *
+ * @deprecated Use CreditCardForm for flag-based rendering or StripeCreditCardForm directly instead.
  */
-function CreditCardForm(props: Props) {
+function LegacyCreditCardForm(props: Props) {
   return (
     <StripeWrapper>
       <CreditCardFormInner {...props} />
@@ -111,7 +113,7 @@ function CreditCardFormInner({
   footerClassName = 'form-actions',
   referrer,
   location,
-  budgetModeText,
+  budgetTerm,
 }: Props) {
   const theme = useTheme();
   const [busy, setBusy] = useState(false);
@@ -236,10 +238,10 @@ function CreditCardFormInner({
           {location !== null && location !== undefined && (
             <FinePrint>
               {tct(
-                'By clicking [buttonText], you authorize Sentry to automatically charge you recurring subscription fees and applicable [budgetModeText] fees. Recurring charges occur at the start of your selected billing cycle for subscription fees and monthly for [budgetModeText] fees. You may cancel your subscription at any time [here:here].',
+                'By clicking [buttonText], you authorize Sentry to automatically charge you recurring subscription fees and applicable [budgetTerm] fees. Recurring charges occur at the start of your selected billing cycle for subscription fees and monthly for [budgetTerm] fees. You may cancel your subscription at any time [here:here].',
                 {
                   buttonText: <b>{buttonText}</b>,
-                  budgetModeText,
+                  budgetTerm,
                   here: (
                     <ExternalLink href="https://sentry.io/settings/billing/cancel/" />
                   ),
@@ -320,4 +322,4 @@ const AlertContent = styled('span')`
   justify-content: space-between;
 `;
 
-export default CreditCardForm;
+export default LegacyCreditCardForm;
