+++ static/gsApp/components/creditCardEdit/stripeSetup.tsx
@@ -1,35 +1,35 @@
 import type {Organization} from 'sentry/types/organization';
 
-import StripeCreditCardForm from 'getsentry/components/stripeForms/stripeCreditCardForm';
+import StripeCreditCardForm from 'getsentry/components/creditCardEdit/stripeForm';
 import type {FTCConsentLocation, Subscription} from 'getsentry/types';
 
 export interface StripeCreditCardSetupProps {
   /**
-   * Handler for cancellation.
+   * budget term to use for fine print.
    */
-  onCancel: () => void;
-  /**
-   * Handler for success.
-   */
-  onSuccess: () => void;
+  budgetTerm: string;
   /**
    * The organization associated with the form
    */
   organization: Organization;
   /**
-   * budget mode text for fine print, if any.
-   */
-  budgetModeText?: string;
-
-  /**
    * Text for the submit button.
    */
   buttonText?: string;
-
   /**
    * Location of form, if any.
    */
   location?: FTCConsentLocation;
+
+  /**
+   * Handler for cancellation.
+   */
+  onCancel?: () => void;
+
+  /**
+   * Handler for success.
+   */
+  onSuccess?: () => void;
   /**
    * Handler for success called with new subscription state.
    */
