+++ static/gsApp/views/amCheckout/steps/addPaymentMethod.tsx
@@ -12,7 +12,7 @@ import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import TextBlock from 'sentry/views/settings/components/text/textBlock';
 
-import CreditCardSetup from 'getsentry/components/creditCardSetup';
+import CreditCardSetup from 'getsentry/components/creditCardEdit/setup';
 import SubscriptionStore from 'getsentry/stores/subscriptionStore';
 import {FTCConsentLocation} from 'getsentry/types';
 import StepHeader from 'getsentry/views/amCheckout/steps/stepHeader';
@@ -45,10 +45,10 @@ function AddPaymentMethod({
     const cardComponent = (
       <AddCardSetup
         organization={organization}
-        onSuccess={onPaymentAccepted}
+        onSuccessWithSubscription={onPaymentAccepted}
         buttonText="Continue"
         location={FTCConsentLocation.CHECKOUT}
-        budgetModeText={subscription.planDetails.budgetTerm}
+        budgetTerm={subscription.planDetails.budgetTerm}
       />
     );
